# perf(ext/http): optimize `set_response` for small responses